+++ src/elements/button/button-static/button-static.component.ts
@@ -9,7 +9,8 @@ import { buttonCommonStyle, buttonPrimaryStyle, SbbButtonCommonElementMixin } fr
  * It displays a static button enhanced with the SBB Design in the 'primary' variant.
  *
  * @slot - Use the unnamed slot to add content to the button-static.
- * @slot icon - Slot used to display the icon, if one is set
+ * @slot icon - Slot used to display the icon, if one is set.
+ * @cssprop [--sbb-button-loading-delay=300ms] - The delay before the loading animation starts, when setting the button into loading state.
  */
 export
 @customElement('sbb-button-static')