+++ src/elements/dialog/dialog-title/__snapshots__/dialog-title.snapshot.spec.snap.js
@@ -4,7 +4,8 @@ export const snapshots = {};
 snapshots["sbb-dialog-title renders Light DOM"] = 
 `<sbb-dialog-title
   level="2"
-  visual-level="3"
+  slot="title-section"
+  visual-level="4"
 >
   Title
 </sbb-dialog-title>