+++ src/elements/radio-button/radio-button-panel/radio-button-panel.component.ts
@@ -70,11 +70,13 @@ class SbbRadioButtonPanelElement<T = string> extends SbbPanelMixin(
   @property()
   public accessor value: T | null = null;
 
-  private _hasSelectionExpansionPanelElement: boolean = false;
+  private _hasSelectionPanelElement: boolean = false;
 
   public override connectedCallback(): void {
     super.connectedCallback();
-    this._hasSelectionExpansionPanelElement = !!this.closest?.('sbb-selection-expansion-panel');
+    this._hasSelectionPanelElement = !!this.closest?.(
+      'sbb-selection-expansion-panel, sbb-selection-action-panel',
+    );
   }
 
   protected override willUpdate(changedProperties: PropertyValues<this>): void {
@@ -96,22 +98,22 @@ class SbbRadioButtonPanelElement<T = string> extends SbbPanelMixin(
   }
 
   /**
-   * As an exception, panels with an expansion-panel attached are always focusable
+   * As an exception, panels with a selection-panel attached are always focusable
    */
   protected override updateFocusableRadios(): void {
     super.updateFocusableRadios();
     const radios = Array.from(this.associatedRadioButtons ?? []) as SbbRadioButtonPanelElement[];
 
     radios
-      .filter((r) => !r.disabled && r._hasSelectionExpansionPanelElement)
+      .filter((r) => !r.disabled && r._hasSelectionPanelElement)
       .forEach((r) => (r.tabIndex = 0));
   }
 
   /**
-   * As an exception, radio-panels with an expansion-panel attached are not checked automatically when navigating by keyboard
+   * As an exception, radio-panels with a selection-panel attached are not checked automatically when navigating by keyboard
    */
   protected override async navigateByKeyboard(next: SbbRadioButtonPanelElement): Promise<void> {
-    if (!this._hasSelectionExpansionPanelElement) {
+    if (!this._hasSelectionPanelElement) {
       await super.navigateByKeyboard(next);
     } else {
       next.focus();
@@ -152,5 +154,9 @@ declare global {
   interface HTMLElementTagNameMap {
     // eslint-disable-next-line @typescript-eslint/naming-convention
     'sbb-radio-button-panel': SbbRadioButtonPanelElement;
+  }
+
+  interface GlobalEventHandlersEventMap {
+    statechange: CustomEvent<SbbStateChange>;
   }
 }