+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.1.0](https://github.com/sbb-design-systems/lyne-components/compare/v3.0.1...v3.1.0) (2025-07-14)
+
+
+### Features
+
+* **buttons:** introduce loading state ([#3468](https://github.com/sbb-design-systems/lyne-components/issues/3468)) ([fcf3efe](https://github.com/sbb-design-systems/lyne-components/commit/fcf3efe64764e127f3d0b5b664eb2374c43eee82))
+* **sbb-dialog:** provide optional close button ([#3910](https://github.com/sbb-design-systems/lyne-components/issues/3910)) ([105b3cc](https://github.com/sbb-design-systems/lyne-components/commit/105b3cc88e77baa2c210f827bbfdaa8f9c507fbf))
+* **sbb-expansion-panel-header:** add CSS variable for justify-content ([#3913](https://github.com/sbb-design-systems/lyne-components/issues/3913)) ([2697d84](https://github.com/sbb-design-systems/lyne-components/commit/2697d84d754ca4b0a72a47e10a4b94622ff188af)), closes [#3912](https://github.com/sbb-design-systems/lyne-components/issues/3912)
+* **sbb-form-field:** add support for custom controls ([#3906](https://github.com/sbb-design-systems/lyne-components/issues/3906)) ([b16bf90](https://github.com/sbb-design-systems/lyne-components/commit/b16bf902e23024a82be36a39d417a407a6eaf05f))
+* **sbb-header-environment:** initial implementation ([#3897](https://github.com/sbb-design-systems/lyne-components/issues/3897)) ([ff7ee89](https://github.com/sbb-design-systems/lyne-components/commit/ff7ee8980360a6caa8aaa9d3bbeab99b5cc61905))
+* **sbb-mini-button:** add possibility to add a label ([#3893](https://github.com/sbb-design-systems/lyne-components/issues/3893)) ([4f8e868](https://github.com/sbb-design-systems/lyne-components/commit/4f8e8686072865a5ae18faacb891ab58a6ebea09))
+* **sbb-option-hint:** initial implementation ([#3790](https://github.com/sbb-design-systems/lyne-components/issues/3790)) ([c8a302f](https://github.com/sbb-design-systems/lyne-components/commit/c8a302f2daf38929505269ec6c2dd64d9add3f6a))
+* **sbb-popover:** add global configurations for delays ([#3905](https://github.com/sbb-design-systems/lyne-components/issues/3905)) ([c1d3ee1](https://github.com/sbb-design-systems/lyne-components/commit/c1d3ee10292eaa75d601ddf08023c13e15fb1b77))
+* **sbb-selection-action-panel:** initial implementation ([#3873](https://github.com/sbb-design-systems/lyne-components/issues/3873)) ([fa60d25](https://github.com/sbb-design-systems/lyne-components/commit/fa60d25ba9613403ec36c136f5ff20ccb4743063))
+
+
+### Bug Fixes
+
+* avoid live announcer accessing DOM when it is not ready ([#3903](https://github.com/sbb-design-systems/lyne-components/issues/3903)) ([1718c52](https://github.com/sbb-design-systems/lyne-components/commit/1718c5203b567cfb8c57380a6c6158773caa147a)), closes [#3902](https://github.com/sbb-design-systems/lyne-components/issues/3902)
+* **deps:** update dependency lit to v3.3.1 (3.x) ([#3914](https://github.com/sbb-design-systems/lyne-components/issues/3914)) ([3418e4c](https://github.com/sbb-design-systems/lyne-components/commit/3418e4c1992301825e4b4033db6c1bd386cea7c6))
+
+
+### Code Refactoring
+
+* **sbb-menu, sbb-menu-button, sbb-menu-link:** assigned specific role ([#3802](https://github.com/sbb-design-systems/lyne-components/issues/3802)) ([176eb24](https://github.com/sbb-design-systems/lyne-components/commit/176eb24053f74422802e676422316adbbc2fcddd))
+
 ## [3.0.1](https://github.com/sbb-design-systems/lyne-components/compare/v3.0.0...v3.0.1) (2025-07-07)